detector_de_neumonia_app/detector_neumonia/requirements.txt
streamlit
torch
fastai
timm
pillow